+++ lib/src/analyzers/lint_analyzer/rules/rules_list/prefer_const_border_radius/prefer_const_border_radius.dart
@@ -17,7 +17,7 @@ part 'visitor.dart';
 class PreferConstBorderRadiusRule extends FlutterRule {
   static const ruleId = 'prefer-const-border-radius';
   static const _issueMessage =
-      'Prefer using const constructor BorderRadius.all';
+      'Prefer using const constructor BorderRadius.all.';
   static const _replaceComment = 'Replace with const BorderRadius constructor.';
 
   PreferConstBorderRadiusRule([Map<String, Object> config = const {}])